% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.keyword.r
\name{as.keyword}
\alias{as.keyword}
\title{As Keyword}
\usage{
as.keyword(keyword_vector)
}
\arguments{
\item{keyword_vectoro}{A vector of keywords from Bing}
}
\description{
This function formats Bing keywords data by removing special characters and ensuring proper typing.
}
\examples{
as.keyword(keyword_vector)
as.keyword(bing_report$keyword)
}

