% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_raw_keyword_data.r
\name{clean_raw_keyword_data}
\alias{clean_raw_keyword_data}
\title{Clean Raw Bing Keyword Data}
\usage{
clean_raw_keyword_data(data_frame)
}
\arguments{
\item{data_frame}{A data frame of raw data from an Bing Keyword Performance Report}
}
\value{
A data frame with the Bing data in a format consistent with other related data sets.
}
\description{
Takes the raw data from an Bing Keyword Performance Report via an API pull and cleans/formats it
to be more easily readable, understandable and consistent with other data sets.
}
\examples{
data  <- raw_keyword_performance_data(from=20151216, to=20151219)
clean_raw_keyword_data <- clean_raw_keyword_data(data)
}

