% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keywords_campaign_device_matchtype_report.r
\name{keywords_campaign_device_matchtype_report}
\alias{keywords_campaign_device_matchtype_report}
\title{Keywords, Campaign, Device, Matchtype Report}
\usage{
keywords_campaign_device_matchtype_report(keywords_elog,
  keyword_filter = NULL, campaign_filter = NULL, write_csv = TRUE,
  file = "bing_keyword_campaign_device_matchtype_report.csv")
}
\arguments{
\item{keywords_elog}{A data frame containing keywords and transaction event log data.}

\item{keyword_filter}{Optional. (e.g. "paleo meals") If provided, generates a data frame and plot only for the *exact* keyword provided. Default to NULL.}

\item{campaign_filter}{Optional. (e.g. "Paleo Performers") If provided, generates a data frame and plot for campaigns where the string is included in the campaign name. Default to NULL.}

\item{write_csv}{logical. If TRUE, creates a csv of the data frame in the working directory with the filename. If false, creates no file. Defaults to \code{TRUE}.}

\item{file}{String. The filename for the generated report if \code{write_csv} is set to \code{TRUE}. Defaults to bing_keyword_campaign_device_matchtype_report.csv".}
}
\value{
The data frame used to create the report.
}
\description{
Take in an bing keyword event log and creates a report
grouping metrics by Keyword, Campaign, Device and Matchtype.
}
\examples{
keywords_campaign_device_matchtype_report(keywords_elog, keyword_filter="paleo meals", campaign_filter="Paleo Performers")
}

