% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keywords_with_earnings.r
\name{keywords_with_earnings}
\alias{keywords_with_earnings}
\title{Keywords with Earnings}
\usage{
keywords_with_earnings(keywords_elog, campaign_filter = NULL,
  write_csv = TRUE, file = "bing_keywords_with_earnings.csv")
}
\arguments{
\item{keywords_elog}{A data frame containing keywords and transaction event log data.}

\item{campaign_filter}{Optional. (e.g. "Paleo Performers") If provided, generates a data frame and plot for campaigns where the string is included in the campaign name. Default to NULL.}

\item{write_csv}{logical. If TRUE, creates a csv of the data frame in the working directory with the filename. If false, creates no file. Defaults to \code{TRUE}.}

\item{file}{String. The filename for the generated report if \code{write_csv} is set to \code{TRUE}. Defaults to adwords_keyword_campaign_device_matchtype_report.csv".}
}
\value{
The data frame used to create the report.
}
\description{
Take in a Bing keyword event log and creates a data frame
containing only those keywords that have earnings.
}
\examples{
keywords_with_earnings(keywords_elog, keyword_filter="paleo meals", campaign_filter="Paleo Performers")
}

