% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_keyword_performance_data.r
\name{raw_keyword_performance_data}
\alias{raw_keyword_performance_data}
\title{Get Raw Keyword Performance Data}
\usage{
raw_keyword_performance_data(from = Sys.Date(), to = Sys.Date(),
  keyword_performance_file_name = "bing_keyword_performance_report.csv")
}
\arguments{
\item{from}{Date object for start date in format <yyyymmdd>. Inclusive.  If left blank, defaults to Sys.Date()}

\item{to}{Date object for end date in format <yyyymmdd>. Inclusive. If left blank, defaults to Sys.Date()}

\item{keyword_performance_file_name}{Filename from which to pull the data.  Bing has a terrible API, so the file must be manually downloaded from https://reporting.bingads.microsoft.com until we figure out the API.}
}
\value{
A data frame with the keyword performance data in the timeframe.
}
\description{
This is a utility function for that pulls a Keyword Performance Report
with the data needed for our analysis from a Bing CSV.
}
\examples{
raw_keyword_performance_data(from=20151216, to=20151219)
}

